Move all subtitle conversion code into subtitles.py
